Move all const-related files to new const directory
