+++ picard/const/attributes.py
