+++ picard/const/locales.py
