+++ picard/coverartarchive.py
@@ -17,7 +17,7 @@
 # along with this program; if not, write to the Free Software
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 
-from picard.attributes import MB_ATTRIBUTES
+from picard.const import MB_ATTRIBUTES
 from picard.i18n import ugettext_attr
 
 # list of types from http://musicbrainz.org/doc/Cover_Art/Types
